+++ cyberpet/entities.cpp
@@ -1,14 +1,14 @@
 #include "entities.h"
 
-int setup_player_entity(unsigned int entities[ENTITY_SIZE][ENTITY_ATTRS], unsigned int player_level) {
+int setup_player_entity(uint64_t entities[ENTITY_SIZE][ENTITY_ATTRS], unsigned int player_level) {
   int entity_id = ENTITY_ID_PLAYER;
   int equipment_id = 20; // 7/20 hp testing
   int equipment_tier = 0;
 
-  entities[entity_id][ENTITY_STR] = get_equip_stat(equipment_id, STAT_STR, equipment_tier);
-  entities[entity_id][ENTITY_DEX] = get_equip_stat(equipment_id, STAT_DEX, equipment_tier);
-  entities[entity_id][ENTITY_INT] = get_equip_stat(equipment_id, STAT_INT, equipment_tier);
-  entities[entity_id][ENTITY_VIT] = get_equip_stat(equipment_id, STAT_VIT, equipment_tier);
+  entities[entity_id][ENTITY_STR] = (uint64_t)get_equip_stat(equipment_id, STAT_STR, equipment_tier);
+  entities[entity_id][ENTITY_DEX] = (uint64_t)get_equip_stat(equipment_id, STAT_DEX, equipment_tier);
+  entities[entity_id][ENTITY_INT] = (uint64_t)get_equip_stat(equipment_id, STAT_INT, equipment_tier);
+  entities[entity_id][ENTITY_VIT] = (uint64_t)get_equip_stat(equipment_id, STAT_VIT, equipment_tier);
   entities[entity_id][ENTITY_LEVEL] = player_level;
 
   entities[entity_id][ENTITY_ALIVE] = 1;
@@ -33,7 +33,7 @@ int setup_player_entity(unsigned int entities[ENTITY_SIZE][ENTITY_ATTRS], unsign
 }
 
 int setup_room_entities(
-  unsigned int entities[ENTITY_SIZE][ENTITY_ATTRS],
+  uint64_t entities[ENTITY_SIZE][ENTITY_ATTRS],
   byte room_wallmap[MAP_W][MAP_H],
   int room_exit_navmap[MAP_W][MAP_H],
   int room_entity_navmap[MAP_W][MAP_H],
@@ -82,7 +82,7 @@ int setup_room_entities(
 }
 
 int setup_entity(
-  unsigned int entities[ENTITY_SIZE][ENTITY_ATTRS],
+  uint64_t entities[ENTITY_SIZE][ENTITY_ATTRS],
   int entity_id,
   int entity_x,
   int entity_y,
@@ -96,11 +96,11 @@ int setup_entity(
   bool is_elite = false; // +3 levels, gold+loot
   bool is_boss = false; // +5 levels, gold+loot
 
-  entities[entity_id][ENTITY_LEVEL] = (unsigned int)world_tile_data[TILE_LEVEL];
-  entities[entity_id][ENTITY_STR] = get_equip_stat(equipment_id, STAT_STR);
-  entities[entity_id][ENTITY_DEX] = get_equip_stat(equipment_id, STAT_DEX);
-  entities[entity_id][ENTITY_INT] = get_equip_stat(equipment_id, STAT_INT);
-  entities[entity_id][ENTITY_VIT] = get_equip_stat(equipment_id, STAT_VIT);
+  entities[entity_id][ENTITY_LEVEL] = (uint64_t)world_tile_data[TILE_LEVEL];
+  entities[entity_id][ENTITY_STR] = (uint64_t)get_equip_stat(equipment_id, STAT_STR);
+  entities[entity_id][ENTITY_DEX] = (uint64_t)get_equip_stat(equipment_id, STAT_DEX);
+  entities[entity_id][ENTITY_INT] = (uint64_t)get_equip_stat(equipment_id, STAT_INT);
+  entities[entity_id][ENTITY_VIT] = (uint64_t)get_equip_stat(equipment_id, STAT_VIT);
 
   entities[entity_id][ENTITY_ALIVE] = 1;
   entities[entity_id][ENTITY_BLESSED_TICKS] = 0;
@@ -123,22 +123,22 @@ int setup_entity(
   return 1;
 }
 
-unsigned int get_entity_stat(unsigned int entities[ENTITY_SIZE][ENTITY_ATTRS], int entity_id, byte stat) {
-  float stat_growth = statpoint_growth_enemy;
+uint64_t get_entity_stat(uint64_t entities[ENTITY_SIZE][ENTITY_ATTRS], int entity_id, byte stat) {
+  double stat_growth = statpoint_growth_enemy;
   if (entity_id == ENTITY_ID_PLAYER) {
     stat_growth = statpoint_growth_player;
   }
 
-  float level = (float)entities[entity_id][ENTITY_LEVEL];
-  float stat_value = (float)entities[entity_id][stat];
-  return (unsigned int)(stat_value * pow(stat_growth, level));
+  double level = (double)entities[entity_id][ENTITY_LEVEL];
+  uint64_t stat_value = entities[entity_id][stat];
+  return stat_value * (uint64_t)pow(stat_growth, level);
 }
 
-int update_entity_stats(unsigned int entities[ENTITY_SIZE][ENTITY_ATTRS], int entity_id) {
-  unsigned int stat_str = get_entity_stat(entities, entity_id, STAT_STR);
-  unsigned int stat_dex = get_entity_stat(entities, entity_id, STAT_DEX);
-  unsigned int stat_int = get_entity_stat(entities, entity_id, STAT_INT);
-  unsigned int stat_vit = get_entity_stat(entities, entity_id, STAT_VIT);
+int update_entity_stats(uint64_t entities[ENTITY_SIZE][ENTITY_ATTRS], int entity_id) {
+  uint64_t stat_str = get_entity_stat(entities, entity_id, STAT_STR);
+  uint64_t stat_dex = get_entity_stat(entities, entity_id, STAT_DEX);
+  uint64_t stat_int = get_entity_stat(entities, entity_id, STAT_INT);
+  uint64_t stat_vit = get_entity_stat(entities, entity_id, STAT_VIT);
 
   entities[entity_id][ENTITY_HP] = get_max_hp(stat_str, stat_dex, stat_int, stat_vit);
   entities[entity_id][ENTITY_SP] = get_max_sp(stat_str, stat_dex, stat_int, stat_vit);
@@ -146,16 +146,16 @@ int update_entity_stats(unsigned int entities[ENTITY_SIZE][ENTITY_ATTRS], int en
   return 1;
 }
 
-int get_main_stat(unsigned int stat_str, unsigned int stat_dex, unsigned int stat_int) {
-  unsigned int stats[] = {stat_str, stat_dex, stat_int};
+byte get_main_stat(uint64_t stat_str, uint64_t stat_dex, uint64_t stat_int) {
+  uint64_t stats[] = {stat_str, stat_dex, stat_int};
   return get_main_stat(stats);
 }
 
-int get_main_stat(unsigned int stats[4]) {
-  unsigned int highest_stat_value = -1;
-  int highest_stat = -1;
+byte get_main_stat(uint64_t stats[4]) {
+  uint64_t highest_stat_value = -1;
+  byte highest_stat = -1;
 
-  for (int stat = 0; stat < 3; stat++) {
+  for (byte stat = 0; stat < 3; stat++) {
     if (stats[stat] > highest_stat_value) {
       highest_stat_value = stats[stat];
       highest_stat = stat;
@@ -165,26 +165,26 @@ int get_main_stat(unsigned int stats[4]) {
   return highest_stat;
 }
 
-unsigned get_entity_max_hp(unsigned int entities[ENTITY_SIZE][ENTITY_ATTRS], int entity_id) {
-  unsigned int stat_str = get_entity_stat(entities, entity_id, ENTITY_STR);
-  unsigned int stat_dex = get_entity_stat(entities, entity_id, ENTITY_DEX);
-  unsigned int stat_int = get_entity_stat(entities, entity_id, ENTITY_INT);
-  unsigned int stat_vit = get_entity_stat(entities, entity_id, ENTITY_VIT);
+uint64_t get_entity_max_hp(uint64_t entities[ENTITY_SIZE][ENTITY_ATTRS], int entity_id) {
+  uint64_t stat_str = get_entity_stat(entities, entity_id, ENTITY_STR);
+  uint64_t stat_dex = get_entity_stat(entities, entity_id, ENTITY_DEX);
+  uint64_t stat_int = get_entity_stat(entities, entity_id, ENTITY_INT);
+  uint64_t stat_vit = get_entity_stat(entities, entity_id, ENTITY_VIT);
 
   return get_max_hp(stat_str, stat_dex, stat_int, stat_vit);
 }
 
-unsigned get_entity_max_sp(unsigned int entities[ENTITY_SIZE][ENTITY_ATTRS], int entity_id) {
-  unsigned int stat_str = get_entity_stat(entities, entity_id, ENTITY_STR);
-  unsigned int stat_dex = get_entity_stat(entities, entity_id, ENTITY_DEX);
-  unsigned int stat_int = get_entity_stat(entities, entity_id, ENTITY_INT);
-  unsigned int stat_vit = get_entity_stat(entities, entity_id, ENTITY_VIT);
+uint64_t get_entity_max_sp(uint64_t entities[ENTITY_SIZE][ENTITY_ATTRS], int entity_id) {
+  uint64_t stat_str = get_entity_stat(entities, entity_id, ENTITY_STR);
+  uint64_t stat_dex = get_entity_stat(entities, entity_id, ENTITY_DEX);
+  uint64_t stat_int = get_entity_stat(entities, entity_id, ENTITY_INT);
+  uint64_t stat_vit = get_entity_stat(entities, entity_id, ENTITY_VIT);
 
   return get_max_sp(stat_str, stat_dex, stat_int, stat_vit);
 }
 
-unsigned int get_max_hp(unsigned int stat_str, unsigned int stat_dex, unsigned int stat_int, unsigned int stat_vit) {
-  int main_stat = get_main_stat(stat_str, stat_dex, stat_int);
+uint64_t get_max_hp(uint64_t stat_str, uint64_t stat_dex, uint64_t stat_int, uint64_t stat_vit) {
+  byte main_stat = get_main_stat(stat_str, stat_dex, stat_int);
 
   switch(main_stat) {
     case STAT_STR:
@@ -201,8 +201,8 @@ unsigned int get_max_hp(unsigned int stat_str, unsigned int stat_dex, unsigned i
   return hp_base + (stat_vit * 2);
 }
 
-unsigned int get_max_sp(unsigned int stat_str, unsigned int stat_dex, unsigned int stat_int, unsigned int stat_vit) {
-  int main_stat = get_main_stat(stat_str, stat_dex, stat_int);
+uint64_t get_max_sp(uint64_t stat_str, uint64_t stat_dex, uint64_t stat_int, uint64_t stat_vit) {
+  byte main_stat = get_main_stat(stat_str, stat_dex, stat_int);
 
   switch(main_stat) {
     case STAT_STR:
